requirements.txt
aiogram==3.2.0
aiohttp==3.9.0
urllib3==2.1.0
python-dotenv==1.0.0
Django>=3.0
psycopg2-binary
jsonfield
sqlalchemy
aiogram-dialog~=1.0
